+++ src/main/java/edu/ucsb/cs156/example/entities/MenuItemReview.java
@@ -23,7 +23,7 @@ public class MenuItemReview {
 
     private long itemId; // unique id for each menu item
     private String reviewerEmail; // email of the reviewer
-    private short stars; // number of stars given to the menu item (0 to 5)
+    private int stars; // number of stars given to the menu item (0 to 5)
     private LocalDateTime dateReviewed; // date the review was written
     private String comment; // comment written by the reviewer
 }
